+++ frontends/main/src/app-pages/VideoPlaylistCollectionPage/VideoPlaylistCollectionPage.tsx
@@ -9,12 +9,16 @@ import {
   learningResourceQueries,
   videoPlaylistQueries,
 } from "api/hooks/learningResources"
+import { formatDurationHuman } from "ol-utilities"
+import { getVideoPlaylistType } from "@/common/utils"
 import type { VideoResource, VideoPlaylistResource } from "api/v1"
 import { ResourceTypeEnum, VideoResourceResourceTypeEnum } from "api/v1"
+import { EpisodeItem } from "./SeriesVideoList"
 import VideoPageHeader from "./VideoPageHeader"
 import FeaturedVideo from "./FeaturedVideo"
 import VideoCollection from "./VideoCollection"
 import RelatedPlaylist from "./RelatedPlaylist"
+import VideoContainer from "./VideoContainer"
 import { FeatureFlags } from "@/common/feature_flags"
 import { useFeatureFlagsLoaded } from "@/common/useFeatureFlagsLoaded"
 
@@ -23,6 +27,20 @@ const Page = styled.div(({ theme }) => ({
   minHeight: "100vh",
 }))
 
+const EpisodeListSection = styled.section(({ theme }) => ({
+  backgroundColor: theme.custom.colors.lightGray1,
+}))
+
+const EpisodeListContainer = styled(VideoContainer)(({ theme }) => ({
+  [theme.breakpoints.up("sm")]: {
+    padding: "0 48px !important",
+  },
+}))
+const EpisodeListUl = styled.ul({
+  listStyle: "none",
+  margin: 0,
+  padding: 0,
+})
 type VideoPlaylistCollectionPageProps = {
   playlistId: number
 }
@@ -71,6 +89,23 @@ const VideoPlaylistCollectionPage: React.FC<
       item.resource_type === VideoResourceResourceTypeEnum.Video,
   )
   const collectionVideos = videos.slice(1)
+  const playlistType = getVideoPlaylistType(playlist)
+
+  const totalVideos = videos.length
+  const totalVideoSeconds = videos.reduce((acc, video) => {
+    const duration = video.video?.duration ?? ""
+    const match = /PT(?:(\d+)H)?(?:(\d+)M)?(?:(\d+)S)?/.exec(duration)
+    if (!match) return acc
+    return (
+      acc +
+      parseInt(match[1] ?? "0") * 3600 +
+      parseInt(match[2] ?? "0") * 60 +
+      parseInt(match[3] ?? "0")
+    )
+  }, 0)
+  const totalTime = formatDurationHuman(
+    `PT${Math.floor(totalVideoSeconds / 3600)}H${Math.floor((totalVideoSeconds % 3600) / 60)}M${totalVideoSeconds % 60}S`,
+  )
 
   const otherCollections = ((similarData ?? []) as VideoPlaylistResource[])
     .filter((resource) => resource.id !== playlistId)
@@ -78,20 +113,43 @@ const VideoPlaylistCollectionPage: React.FC<
 
   return (
     <Page>
-      <VideoPageHeader playlist={playlist} />
+      <VideoPageHeader playlist={playlist} isSeries={playlistType} />
 
       {isLoading ? (
         <Skeleton variant="rectangular" width="100%" height={460} />
       ) : videos.length > 0 ? (
-        <FeaturedVideo video={videos[0]} href={getVideoHref(videos[0])} />
+        <FeaturedVideo
+          video={videos[0]}
+          href={getVideoHref(videos[0])}
+          isSeries={playlistType}
+          totalVideos={totalVideos}
+          totalTime={totalTime}
+        />
       ) : null}
-
-      <VideoCollection
-        videos={collectionVideos}
-        isLoading={isLoading}
-        getHref={getVideoHref}
-      />
-
+      {!playlistType && (
+        <VideoCollection
+          videos={collectionVideos}
+          isLoading={isLoading}
+          getHref={getVideoHref}
+        />
+      )}
+      {playlistType && (
+        <EpisodeListSection>
+          <EpisodeListContainer>
+            <EpisodeListUl>
+              {videos.map((video, i) => (
+                <EpisodeItem
+                  key={video.id}
+                  episode={video}
+                  index={i + 1}
+                  onPlayClick={() => getVideoHref(video)}
+                  isPlayable
+                />
+              ))}
+            </EpisodeListUl>
+          </EpisodeListContainer>
+        </EpisodeListSection>
+      )}
       <RelatedPlaylist
         collections={otherCollections}
         isLoading={similarLoading}
